LPS-55110 Move test to JournalArticleServiceTest
