+++ portal-impl/test/integration/com/liferay/portlet/dynamicdatamapping/service/DDMTemplateServiceTest.java
@@ -28,16 +28,12 @@ import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.MainServletTestRule;
 import com.liferay.portal.util.PortalUtil;
 import com.liferay.portlet.asset.model.AssetEntry;
-import com.liferay.portlet.dynamicdatamapping.RequiredTemplateException;
 import com.liferay.portlet.dynamicdatamapping.TemplateDuplicateTemplateKeyException;
 import com.liferay.portlet.dynamicdatamapping.TemplateNameException;
 import com.liferay.portlet.dynamicdatamapping.TemplateScriptException;
-import com.liferay.portlet.dynamicdatamapping.model.DDMStructure;
 import com.liferay.portlet.dynamicdatamapping.model.DDMTemplate;
 import com.liferay.portlet.dynamicdatamapping.model.DDMTemplateConstants;
 import com.liferay.portlet.dynamicdatamapping.util.comparator.TemplateIdComparator;
-import com.liferay.portlet.journal.model.JournalArticle;
-import com.liferay.portlet.journal.util.test.JournalTestUtil;
 
 import java.util.List;
 
@@ -149,31 +145,6 @@ public class DDMTemplateServiceTest extends BaseDDMServiceTestCase {
 		Assert.assertNull(
 			DDMTemplateLocalServiceUtil.fetchDDMTemplate(
 				template.getTemplateId()));
-	}
-
-	@Test
-	public void testDeleteTemplateReferencedByJournalArticles()
-		throws Exception {
-
-		DDMStructure structure = addStructure(
-			PortalUtil.getClassNameId(JournalArticle.class.getName()),
-			"Test Structure");
-
-		DDMTemplate template = addDisplayTemplate(
-			structure.getPrimaryKey(), "Test Display Template");
-
-		JournalTestUtil.addArticleWithXMLContent(
-			group.getGroupId(), "<title>Test Article</title>",
-			structure.getStructureKey(), template.getTemplateKey());
-
-		try {
-			DDMTemplateLocalServiceUtil.deleteTemplate(
-				template.getTemplateId());
-
-			Assert.fail();
-		}
-		catch (RequiredTemplateException rse) {
-		}
 	}
 
 	@Test
